+++ src/lib/components/side-menu/items/layers/LayerItem.tsx
@@ -1,49 +1,45 @@
 import { useState } from 'react'
-import { useDocumentContext } from '../../../../../hooks/useDocumentContext'
 import { MdDelete, MdModeEdit } from 'react-icons/md'
 import { FaCheckCircle } from 'react-icons/fa'
 import { IoMdRemoveCircle } from 'react-icons/io'
+import { useDocumentContext } from '../../../hooks/useDocumentContext'
+import { useTranslation } from 'react-i18next'
 
 interface IEditAnotationParams {
   group: { id: string; name: string }
-  setVisible: (visible: boolean) => void
   update: (id: string, name: string) => void
   remove: (id: string) => void
-  choosing: boolean
 }
-const EditAnotation = ({
-  group,
-  setVisible,
-  update,
-  remove,
-  choosing,
-}: IEditAnotationParams) => {
-  const { setEditGroupId, setGroupId } = useDocumentContext()
+const LayerItem = ({ group, update, remove }: IEditAnotationParams) => {
+  const { setGroupId, setIsEditMode, groupId } = useDocumentContext()
+  const { t } = useTranslation()
   const [input, setInput] = useState<string>(group.name)
   const [disabled, setDisabled] = useState<boolean>(true)
 
   return (
     <button
-      className="efw-flex efw-items-center efw-p-4 efw-rounded-md efw-bg-gray-light dark:efw-bg-gray-dark-medium hover:efw-bg-opacity-50 dark:hover:efw-bg-opacity-50"
+      className={`efw-rounded-md efw-bg-gray-light dark:efw-bg-gray-dark-medium efw-border-2 ${
+        group.id === groupId
+          ? 'efw-border-black dark:efw-border-white'
+          : 'efw-border-transparent'
+      } hover:efw-bg-opacity-50 dark:hover:efw-bg-opacity-50`}
       onClick={() => {
         if (!disabled) return
-        if (choosing) setGroupId(group.id)
-        else setEditGroupId(group.id)
-        setVisible(false)
+        setGroupId(group.id)
       }}
     >
-      <input
-        className="efw-bg-transparent efw-p-1 efw-border-b-2 disabled:efw-border-transparent disabled:efw-pointer-events-none efw-border-white efw-outline-none"
-        name={group.id}
-        onClick={(e) => e.stopPropagation()}
-        type="text"
-        value={input}
-        onChange={(e) => setInput(e.target.value)}
-        onKeyDown={(e) => e.stopPropagation()}
-        disabled={disabled}
-      />
-      <span className="efw-flex-1"></span>
-      {!choosing && (
+      <div className="efw-flex efw-p-4 efw-pb-2 efw-items-center efw-gap-2">
+        <input
+          className="efw-bg-transparent efw-w-full efw-p-1 efw-border-b-2 disabled:efw-border-transparent efw-uppercase efw-font-bold disabled:efw-pointer-events-none efw-border-white efw-outline-none"
+          name={group.id}
+          onClick={(e) => e.stopPropagation()}
+          type="text"
+          value={input}
+          onChange={(e) => setInput(e.target.value)}
+          onKeyDown={(e) => e.stopPropagation()}
+          disabled={disabled}
+        />
+
         <div
           className="efw-flex efw-h-full efw-items-center efw-gap-2"
           onClick={(e) => e.stopPropagation()}
@@ -81,9 +77,20 @@ const EditAnotation = ({
             </>
           )}
         </div>
-      )}
+      </div>
+      <div
+        className="efw-w-full efw-outline-none efw-text-center efw-uppercase efw-text-[12px] efw-py-2"
+        onClick={(e) => {
+          e.stopPropagation()
+          if (!disabled) return
+          setGroupId(group.id)
+          setIsEditMode(true)
+        }}
+      >
+        {t('editLayer')}
+      </div>
     </button>
   )
 }
 
-export default EditAnotation
+export default LayerItem